Move the character nav out
